+++ code/modules/hydroponics/grown/banana.dm
@@ -23,21 +23,19 @@
 	trash = /obj/item/grown/bananapeel
 	filling_color = "#FFFF00"
 	bitesize = 5
-	distill_reagent = "bananahonk"
-	tastes = list("banana" = 1)
 
 /obj/item/reagent_containers/food/snacks/grown/banana/suicide_act(mob/user)
-	user.visible_message("<span class='suicide'>[user] is aiming [src] at [user.p_them()]self! It looks like [user.p_theyre()] trying to commit suicide.</span>")
+	user.visible_message("<span class='suicide'>[user] is aiming the [src.name] at themself! It looks like \he's trying to commit suicide.</span>")
 	playsound(loc, 'sound/items/bikehorn.ogg', 50, 1, -1)
 	sleep(25)
 	if(!user)
-		return OXYLOSS
+		return (OXYLOSS)
 	user.say("BANG!")
 	sleep(25)
 	if(!user)
-		return OXYLOSS
-	user.visible_message("<B>[user]</B> laughs so hard [user.p_they()] begin[user.p_s()] to suffocate!")
-	return OXYLOSS
+		return (OXYLOSS)
+	user.visible_message("<B>[user]</B> laughs so hard they begin to suffocate!")
+	return (OXYLOSS)
 
 /obj/item/grown/bananapeel
 	seed = /obj/item/seeds/banana
@@ -50,17 +48,10 @@
 	throw_speed = 3
 	throw_range = 7
 
-/obj/item/grown/bananapeel/Initialize(mapload)
-	. = ..()
-	choose_icon_state()
-
-/obj/item/grown/bananapeel/proc/choose_icon_state()
-	icon_state = "[icon_state]_[rand(1, 3)]"
-
 /obj/item/grown/bananapeel/suicide_act(mob/user)
-	user.visible_message("<span class='suicide'>[user] is deliberately slipping on [src]! It looks like [user.p_theyre()] trying to commit suicide.</span>")
+	user.visible_message("<span class='suicide'>[user] is deliberately slipping on the [src.name]! It looks like \he's trying to commit suicide.</span>")
 	playsound(loc, 'sound/misc/slip.ogg', 50, 1, -1)
-	return BRUTELOSS
+	return (BRUTELOSS)
 
 
 // Mimana - invisible sprites are totally a feature!
@@ -83,7 +74,6 @@
 	icon_state = "mimana"
 	trash = /obj/item/grown/bananapeel/mimanapeel
 	filling_color = "#FFFFEE"
-	distill_reagent = "silencer"
 
 /obj/item/grown/bananapeel/mimanapeel
 	seed = /obj/item/seeds/banana/mime
@@ -108,27 +98,27 @@
 /obj/item/reagent_containers/food/snacks/grown/banana/bluespace
 	seed = /obj/item/seeds/banana/bluespace
 	name = "bluespace banana"
-	icon_state = "bluenana"
+	icon_state = "banana_blue"
 	trash = /obj/item/grown/bananapeel/bluespace
 	filling_color = "#0000FF"
 	origin_tech = "biotech=3;bluespace=5"
-	wine_power = 0.6
-	wine_flavor = "slippery hypercubes"
 
 /obj/item/grown/bananapeel/bluespace
 	seed = /obj/item/seeds/banana/bluespace
 	name = "bluespace banana peel"
 	desc = "A peel from a bluespace banana."
-	icon_state = "bluenana_peel"
+	icon_state = "banana_peel_blue"
 
 // Other
 /obj/item/grown/bananapeel/specialpeel     //used by /obj/item/clothing/shoes/clown_shoes/banana_shoes
 	name = "synthesized banana peel"
 	desc = "A synthetic banana peel."
+	trip_stun = 2
+	trip_weaken = 2
+	trip_chance = 100
+	trip_walksafe = FALSE
+	trip_verb = TV_SLIP
 
-/obj/item/grown/bananapeel/specialpeel/ComponentInitialize()
-	AddComponent(/datum/component/slippery, src, 2, 2, 100, 0, FALSE)
-
-/obj/item/grown/bananapeel/specialpeel/after_slip(mob/living/carbon/human/H)
-	. = ..()
-	qdel(src)
+/obj/item/grown/bananapeel/specialpeel/on_trip(mob/living/carbon/human/H)
+	if(..())
+		qdel(src)
