+++ code/modules/mob/living/silicon/robot/death.dm
@@ -1,6 +1,4 @@
 /mob/living/silicon/robot/gib()
-	if(!death(TRUE) && stat != DEAD)
-		return FALSE
 	//robots don't die when gibbed. instead they drop their MMI'd brain
 	var/atom/movable/overlay/animation = null
 	notransform = 1
@@ -18,50 +16,41 @@
 	flick("gibbed-r", animation)
 	robogibs(loc)
 
-	GLOB.alive_mob_list -= src
-	GLOB.dead_mob_list -= src
-	QDEL_IN(animation, 15)
-	QDEL_IN(src, 15)
-	return TRUE
+	living_mob_list -= src
+	dead_mob_list -= src
+	spawn(15)
+		if(animation)	qdel(animation)
+		if(src)			qdel(src)
 
 /mob/living/silicon/robot/dust()
-	if(!death(TRUE) && stat != DEAD)
-		return FALSE
+	death(1)
+	var/atom/movable/overlay/animation = null
 	notransform = 1
 	canmove = 0
+	icon = null
 	invisibility = 101
-	dust_animation()
-	if(mmi)
-		qdel(mmi)	//Delete the MMI first so that it won't go popping out.
-	GLOB.dead_mob_list -= src
-	QDEL_IN(src, 15)
-	return TRUE
 
-/mob/living/silicon/robot/dust_animation()
-	var/atom/movable/overlay/animation = null
 	animation = new(loc)
 	animation.icon_state = "blank"
 	animation.icon = 'icons/mob/mob.dmi'
 	animation.master = src
+
 	flick("dust-r", animation)
 	new /obj/effect/decal/remains/robot(loc)
-	QDEL_IN(animation, 15)
+	if(mmi)		qdel(mmi)	//Delete the MMI first so that it won't go popping out.
+
+	dead_mob_list -= src
+	spawn(15)
+		if(animation)	qdel(animation)
+		if(src)			qdel(src)
+
 
 /mob/living/silicon/robot/death(gibbed)
-	if(can_die())
-		if(!gibbed && deathgasp_on_death)
-			emote("deathgasp", force = TRUE)
-
-		if(module)
-			module.handle_death(src, gibbed)
-
-	// Only execute the below if we successfully died
-	. = ..(gibbed)
-	if(!.)
-		return FALSE
-
-	diag_hud_set_status()
-	diag_hud_set_health()
+	if(stat == DEAD)	return
+	if(!gibbed)
+		emote("deathgasp")
+	stat = DEAD
+	update_canmove()
 	if(camera)
 		camera.status = 0
 	update_headlamp(1) //So borg lights are disabled when killed.
@@ -70,6 +59,13 @@
 		var/obj/machinery/recharge_station/RC = loc
 		RC.go_out()
 
+	sight |= SEE_TURFS|SEE_MOBS|SEE_OBJS
+	see_in_dark = 8
+	see_invisible = SEE_INVISIBLE_LEVEL_TWO
 	update_icons()
+	timeofdeath = world.time
+	if(mind)	mind.store_memory("Time of death: [station_time_timestamp("hh:mm:ss", timeofdeath)]", 0)
 
-	SSblackbox.ReportDeath(src)
+	sql_report_cyborg_death(src)
+
+	return ..(gibbed)
